requirements.txt
azure-functions
requests
pyodbc

